clipboard_lang_shifter/requirements.txt
pyperclip
keyboard
pystray
pillow
plyer